Create FlavIntDataGroup
Simple implementation of a FlavIntData object which works for combinations of
NuFlavIntGroup's. Creates a single level dictionary of NuFlavIntGroup's in
which the data is then contained.

Mechanism added for adding two FlavIntDataGroup's which iterates down to the
lowest level of the dictionary until coincident numpy arrays are found, after
which the appropriate sub-element is made equal to the concatenation of the two
arrays.