Scaled poisson fluctuation (#597)
* handle compound Poisson fluctuations correctly
* warn user about uninitialized errors
* Add documentation about the method in the doc-string.
* infer data type for histogram values

When maps are fluctuated with the "scaled_poisson" method, they take into account
the stored error. If the error is different from Poisson errors (sqrt(N)),
a scaled Poisson process will be used to fluctuate in such a way that
the standard deviation matches the stored standard deviation.

The standard deviation is now always conserved when fluctuating with
the 'scaled_poisson' method.