fix resetting parameter values for 2. octant fit (#710)
* fix resetting parameter values for 2. octant fit

In octant fitting, if parameter values are not reset free between the single fits, the intended behavior is that after fitting the first octant, the parameter values (except for theta_23) are reset to where the first fit started. Due to the missing deepcopy, instead, they were set to the end point of the first fit instead.

* bug fix on conv_llh (#707)

Co-authored-by: Krishnamoorthy Jayakumar <kjayakumar@submit-1.icecube.wisc.edu>

* fix resetting parameter values for 2. octant fit

In octant fitting, if parameter values are not reset free between the single fits, the intended behavior is that after fitting the first octant, the parameter values (except for theta_23) are reset to where the first fit started. Due to the missing deepcopy, instead, they were set to the end point of the first fit instead.

* Cap pint version

Starting at pint 0.20, the property `pint.quantity` no longer exists, which is used in `pisa/utils/comparisons.py`. Quick fix is to cap the version for now.

* Fix order of imports

It turns out that importing pip before setuptools causes an assertion error during the import of setuptools, but only starting at version 60 of setuptools.  That was fun to figure out.

Co-authored-by: Krishnamoorthi Jayakumar <43264985+krish9600@users.noreply.github.com>
Co-authored-by: Krishnamoorthy Jayakumar <kjayakumar@submit-1.icecube.wisc.edu>
Co-authored-by: Elisa Lohfink <elohfink@cobalt08.icecube.wisc.edu>
Co-authored-by: Alex Trettin <alexander.trettin@physik.hu-berlin.de>